refactor(workspace): boot the workspace from its own module
`host.tsx` crossed the file-size cap when the machine hub joined its boot
sequence. The seam was already there: `startWorkspaceHost` wires a daemon, a
hub and a PTY registry together and hands back a root, while everything above
it is the component. They shared only the orchestrator.
